imbot: move Weixin QR provisioning client into the imbot module
The iLink QR client that provisions Weixin bot credentials lived in the
server-layer handler, so the reusable imbot module (CLI + GUI) could not use
it — the same layering issue already fixed for Feishu. Extract the
platform-level HTTP client into imbot/platform/weixin/registration.go as
QRClient, owning the iLink endpoints and response types.

The server handler now holds a *weixin.QRClient and keeps only the HTTP
start/status/cancel session model, rate limiting and credential persistence.
Client tests move to the weixin package; NewQRClient also fixes the
previously nil http.Client in those tests.

This mirrors imbot/platform/feishu/registration.go. No new dependencies —
the extracted client is stdlib-only.
